dev(block_import_refactoring): PreValidate trait
