+++ crates/client/block_import/Cargo.toml
@@ -25,6 +25,7 @@ dp-receipt = { workspace = true }
 dp-state-update = { workspace = true }
 dp-transactions = { workspace = true }
 dp-utils = { workspace = true }
+dp-validation = { workspace = true }
 
 bonsai-trie = { workspace = true }
 starknet-core = { workspace = true }
