+++ crates/client/mempool/Cargo.toml
@@ -34,6 +34,7 @@ dp-receipt.workspace = true
 dp-state-update.workspace = true
 dp-transactions.workspace = true
 dp-utils.workspace = true
+dp-validation.workspace = true
 
 # Starknet
 blockifier.workspace = true
