+++ crates/client/sync/Cargo.toml
@@ -32,6 +32,7 @@ dp-receipt = { workspace = true }
 dp-state-update = { workspace = true }
 dp-transactions = { workspace = true }
 dp-utils = { workspace = true }
+dp-validation = { workspace = true }
 
 # Starknet
 bonsai-trie = { workspace = true }
