+++ crates/primitives/validation/Cargo.toml
@@ -1,0 +1,33 @@
+[package]
+description = "Deoxys primitives for starknet validation logic"
+name = "dp-validation"
+authors.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+version.workspace = true
+homepage.workspace = true
+
+[package.metadata.docs.rs]
+targets = ["x86_64-unknown-linux-gnu"]
+
+[dependencies]
+
+# Deoxys
+dp-chain-config = { workspace = true }
+dp-convert = { workspace = true }
+dp-receipt = { workspace = true }
+dp-transactions = { workspace = true }
+
+# Starknet
+blockifier = { workspace = true }
+starknet-core = { workspace = true }
+starknet-types-core = { workspace = true }
+starknet_api = { workspace = true }
+
+# Other
+lazy_static = { workspace = true }
+primitive-types.workspace = true
+serde = { workspace = true, features = ["derive"] }
+serde_json = { workspace = true }
+thiserror = { workspace = true }
